+++ House.h
@@ -23,14 +23,14 @@ class House {
     //    House(int hId, const string &description, const string &city, int houseRating,
     //           const Date &start, const Date &anEnd, int consumingPoint, int minOccupierRating,
     //           const vector<vector<string>> &reviews);
-    House(int hId, const string &description, const string &city, int houseRating = -11,
+    House(int hId, const string &description, const string &city, double houseRating = -11,
           const Date &start = Date(), const Date &anEnd = Date(), int consumingPoint = 0, int minOccupierRating = 0,
           const vector<vector<string>> &reviews = {});
     const string &getHid() const;
 
     void setHid(const string &hId);
 
-    int getHouseRating() const;
+    double getHouseRating() const;
 
     void setHouseRating(int houseRating);
 
@@ -61,7 +61,7 @@ class House {
 
    private:
     int hID;
-    int houseRating;
+    double houseRating;
     string description;
     string city;
     Date start;
